requirements.txt
requests>=2.28.0
pywin32>=305; sys_platform == 'win32'
python-dotenv>=1.0.0
